test(electrum): fix deserialize coinbase tx

+++ incubator/tbk-electrum-daemon-client/tbk-electrum-daemon-client-core/src/main/java/org/tbk/electrum/model/SimpleTx.java
@@ -28,7 +28,7 @@ public class SimpleTx implements Tx {
         @NonNull
         String txHash;
 
-        int outputIndex;
+        long outputIndex;
 
         long sequenceNumber;
