+++ incubator/tbk-electrum-daemon-client/tbk-electrum-daemon-client-core/src/main/java/org/tbk/electrum/model/Tx.java
@@ -13,7 +13,7 @@ public interface Tx {
     interface TxInput {
         String getTxHash();
 
-        int getOutputIndex();
+        long getOutputIndex();
 
         /**
          * Some electrum rpc responses can contain an address of an input.
